requirements.txt
numpy
Pillow
streamlit
tensorflow
protobuf==3.20